fix(core): run instance middleware after control routes
